rename: input boxes component 분리
